requirements.txt
aioredis
Flask
mysql-connector-python
cryptography